CSS Colored Markers
Understood different color combinations in CSS in much depth.

Working Link : https://csscoloredmarkers.netlify.app/

Understood different color combinations .

•used RGB for Red Color.
•used Hexacode for Green Color.
•used HSL for Blue Color.

•learned about primary ,secondary, and tertiary color combinations.

•Hexadecimal, or base 16 values, go from 0 - 9, then A - F:

0, 1, 2, 3, 4, 5, 6, 7, 8, 9, A, B, C, D, E, F
With hex colors, 00 is 0% of that color, and FF is 100%. So #00FF00 translates to 0% red, 100% green, and 0% blue, and is the same as rgb(0, 255, 0).

•The HSL color model, or hue, saturation, and lightness, is another way to represent colors.

The CSS hsl function accepts 3 values: a number from 0 to 360 for hue, a percentage from 0 to 100 for saturation, and a percentage from 0 to 100 for lightness.